13 - updated Token_Type enum to be more accurate to computer instruction set and hardware, updated keyboard.txt in examples folder

+++ include/structures/file_manager.h
@@ -6,14 +6,14 @@ typedef struct{
 
     char **raw_text;
     FILE *file;
-    int num_lines;
+    uint16_t num_lines;
     char *path;
 
 } File_Info;
 
 typedef struct{
 
-    int amount_inputs;
+    uint16_t amount_inputs;
     File_Info *inputs;
     File_Info output;
